requirements.txt
python-telegram-bot[all]==20.7
yt-dlp==2024.12.13
google-generativeai==0.8.3
aiohttp==3.9.1
tenacity==8.2.3
prometheus-client==0.19.0
python-dotenv==1.0.0 